duplicate cap, handle delete edge cases, and more...
