+++ apps/web/app/(org)/dashboard/caps/components/Folder.tsx
@@ -6,7 +6,9 @@ import { useEffect, useRef, useState } from "react";
 import { toast } from "sonner";
 import Link from "next/link";
 import { useTheme } from "../../Contexts";
-import { deleteFolder, updateFolder, moveVideoToFolder } from "../../folder/[id]/actions";
+import { deleteFolder } from "@/actions/folders/deleteFolder";
+import { updateFolder } from "@/actions/folders/updateFolder";
+import { moveVideoToFolder } from "@/actions/folders/moveVideoToFolder";
 import { registerDropTarget } from "../../folder/[id]/components/ClientCapCard";
 import { ConfirmationDialog } from "../../_components/ConfirmationDialog";
 import { FoldersDropdown } from "./FoldersDropdown";
