+++ apps/web/app/(org)/dashboard/caps/components/NewFolderDialog.tsx
@@ -11,7 +11,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faFolderPlus } from "@fortawesome/free-solid-svg-icons";
 import React, { useEffect, useRef, useState } from "react";
 import clsx from "clsx";
-import { createFolder } from "../../folder/[id]/actions";
+import { createFolder } from "@/actions/folders/createFolder";
 import { toast } from "sonner";
 import { BlueFolder, NormalFolder, RedFolder, YellowFolder } from "./Folders";
