+++ apps/web/app/(org)/dashboard/folder/[id]/components/SubfolderDialog.tsx
@@ -13,7 +13,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faFolderPlus } from "@fortawesome/free-solid-svg-icons";
 import React, { useEffect, useRef, useState } from "react";
 import clsx from "clsx";
-import { createFolder } from "../actions";
+import { createFolder } from "@/actions/folders/createFolder";
 import { toast } from "sonner";
 import { BlueFolder, NormalFolder, RedFolder, YellowFolder } from "../../../caps/components/Folders";
