+++ apps/web/app/(org)/dashboard/folder/[id]/components/useUploadPlaceholders.ts
@@ -1,0 +1,33 @@
+import { useState, useCallback } from "react";
+
+export interface UploadPlaceholder {
+  id: string;
+  progress: number;
+  thumbnail?: string;
+  uploadProgress?: number;
+}
+
+export function useUploadPlaceholders() {
+  const [uploadPlaceholders, setUploadPlaceholders] = useState<UploadPlaceholder[]>([]);
+
+  const handleUploadStart = useCallback((id: string, thumbnail?: string) => {
+    setUploadPlaceholders((prev) => [{ id, progress: 0, thumbnail }, ...prev]);
+  }, []);
+
+  const handleUploadProgress = useCallback((id: string, progress: number, uploadProgress?: number) => {
+    setUploadPlaceholders((prev) =>
+      prev.map((u) => (u.id === id ? { ...u, progress, uploadProgress } : u))
+    );
+  }, []);
+
+  const handleUploadComplete = useCallback((id: string) => {
+    setUploadPlaceholders((prev) => prev.filter((u) => u.id !== id));
+  }, []);
+
+  return {
+    uploadPlaceholders,
+    handleUploadStart,
+    handleUploadProgress,
+    handleUploadComplete,
+  };
+}
